extract "renderScene" into its own file
